+++ kernel/src/interrupts/cont_access.rs
@@ -23,73 +23,117 @@
     Contained within the interrupts module
 */
 
-#![allow(unused)]
+#![allow(dead_code)]
 
+use core::cell::SyncUnsafeCell;
 #[cfg(test)]
-use core::sync::atomic::{AtomicU8, Ordering};
-use core::{cell::SyncUnsafeCell, fmt::Debug};
+use core::sync::atomic::AtomicU8;
+use core::sync::atomic::{AtomicBool, AtomicU32, Ordering};
 #[cfg(not(test))]
 use {crate::PANIC, core::any::type_name};
 
 /// Count of how far down in interrupt handlers we are,
 /// where a zero means that we're not in an interrupt handler.
 #[unsafe(export_name = "int_handler_count")]
-static mut INT_HANDLER_COUNT: u32 = 0;
+static INTERRUPT_DEPTH: AtomicU32 = AtomicU32::new(0);
 
-/// Incremented whenever `get_ref` or `write` would invoke a `badbug`.
+/// Incremented whenever [`ContAccess::check_access`] would fail.
 #[cfg(test)]
 static CONT_ACCESS_PANICS: AtomicU8 = AtomicU8::new(0);
 
-/// Continuous access type.
+/// The continuous access type, (or just CA for short).
 ///
-/// Allows obtain references of and writing to [`Sync`] types without ANY
-/// locking, blocking or failing, so long as the [`ContAccess`] is **NEVER**
-/// accessed by any interrupt handlers (it's only accessed continuously).
+/// Basically just `Cell` and `ExclusiveMap` combined, but without all of the
+/// annoyances they give when you try to actually use them.
 ///
-/// This allows [`ContAccess`]es to be mutated while immutable references to
-/// them are active, as can be seen in the below code example:
+/// The only rules when accessing a CA is to **NOT**
+///
+/// - access a CA inside an interrupt handler,
+/// - or inside a call to it's [`ContAccess::btemap`]
+///
+/// doing so will cause a `badbug` to be triggered and
+/// the kernel to crash horrifically, ruining everyone's day.
 /// ```
-/// let cont: ContAccess<u32> = ContAccess::new(0x42);
-/// let cont_ref: &u32 = cont.get_ref();
-/// cont.write(87); // allowed!
-/// print!("{}", cont_ref); // prints '87'
-/// ```
-struct ContAccess<T: Sync>(SyncUnsafeCell<T>);
+pub struct ContAccess<T> {
+    data: SyncUnsafeCell<T>,
+    /// Set when in btemap, fails check_access
+    locked: AtomicBool,
+}
 
-impl<T: Sync + Debug> ContAccess<T> {
-    /// Creates a new cont access.
-    pub fn new(v: T) -> ContAccess<T> {
-        ContAccess(SyncUnsafeCell::new(v))
+impl<T> ContAccess<T> {
+    /// Creates a new CA from the given `v`.
+    pub const fn new(v: T) -> ContAccess<T> {
+        ContAccess {
+            data: SyncUnsafeCell::new(v),
+            locked: AtomicBool::new(false),
+        }
     }
 
-    /// Returns a reference to the contained value.
+    /// Checks that the CA isn't locked or in an interrupt handler,
+    /// triggering a `badbug` if so.
     ///
-    /// Panics if in an interrupt handler.
-    pub fn get_ref(&self) -> &T {
-        // Safety: Only write is some incs in int handlers
-        let cnt = unsafe { INT_HANDLER_COUNT };
-        if cnt != 0 {
+    /// This means that if this function returns, it's guaranteed that this CA
+    /// will never be accessed from anywhere else (due to CA's being amazing).
+    ///
+    /// Increments [`CONT_ACCESS_PANICS`] instead of triggering
+    /// a `badbug` if in a test build.
+    fn check_access(&self) {
+        let locked = self.locked.load(Ordering::Relaxed);
+        let int_depth = INTERRUPT_DEPTH.load(Ordering::Relaxed);
+
+        // we want to print both locked & depth every fail
+        if locked || int_depth != 0 {
+            #[cfg(not(test))]
+            PANIC!(badbug "ContAccess was accessed in a bad state
+Interrupt depth: {int_depth} {}
+Type: {}", if locked {"Locked"} else {""}, type_name::<T>());
             #[cfg(test)]
             CONT_ACCESS_PANICS.fetch_add(1, Ordering::Relaxed);
-            #[cfg(not(test))]
-            PANIC!(badbug "hit ContAccess:get_ref with a handler count of {cnt}\nAccess type: {}", type_name::<T>())
         }
-        unsafe { &*self.0.get() }
     }
 
-    /// Writes `v` into the contained value.
+    /// Runs the passed function on the contained value, returning it's result.
     ///
-    /// Panics if in an interrupt handler.
-    pub fn write(&self, v: T) {
-        // Safety: Only write is some incs in int handlers
-        let cnt = unsafe { INT_HANDLER_COUNT };
-        if cnt != 0 {
-            #[cfg(test)]
-            CONT_ACCESS_PANICS.fetch_add(1, Ordering::Relaxed);
-            #[cfg(not(test))]
-            PANIC!(badbug "hit ContAccess:write with a handler count of {cnt}\nWrite value: {v:?} Access type: {}", type_name::<T>())
-        }
-        unsafe { *self.0.get() = v }
+    /// Accessing this CA in any other way inside the passed function triggers
+    /// a `badbug`, as can be seen in the below example:
+    ///
+    /// ```
+    /// let ca = ContAccess::new(42);
+    /// ca.btemap(|v|{
+    ///     println!("{v:?}"); // fine...
+    ///     ca.write(15);      // BADBUG TRIGGERED!!
+    ///     println!("{v:?}"); // unreachable
+    /// })
+    ///
+    /// ```
+    #[rustfmt::skip]
+    pub fn btemap<R>(&self, f: impl FnOnce(&mut T) -> R) -> R { // BETTER THAN EXCLUSIVE MAP!!!
+        self.check_access();
+        // check_access ensures that locked is false, and the
+        // nature of ContAccess ensures that it will stay false
+        self.locked.store(true, Ordering::Relaxed);
+
+        // SAFETY: check_access ensures that we have an exclusive access
+        let res = unsafe { f(&mut *self.data.get()) };
+        self.locked.store(false, Ordering::Relaxed);
+        res
+    }
+
+    /// Sets the contained value to `val`.
+    pub fn write(&self, val: T) {
+        self.check_access();
+        // SAFETY: check_access ensures that we have an exclusive access
+        unsafe { *self.data.get() = val }
+    }
+}
+
+impl<T: Copy> ContAccess<T> {
+    /// Copies the contained value then returns it.
+    pub fn copy(&self) -> T {
+        self.check_access();
+        // SAFETY: T: Copy allows copying out of self.data
+        // and check_access ensures no active mutations
+        unsafe { *self.data.get() }
     }
 }
 
@@ -97,31 +141,46 @@ impl<T: Sync + Debug> ContAccess<T> {
 mod tests {
     use super::*;
 
-    /// Tests that [`ContAccess::get_ref`] and [`ContAccess::write`] panic when [`INT_HANDLER_COUNT`] is nonzero.
+    /// Tests that [`ContAccess::check_access`] panics when
+    /// [`INTERRUPT_DEPTH`] is nonzero.
     #[test_case]
-    fn cont_access_panics() {
-        let cont = ContAccess::new(0x42);
+    fn badbug_on_ints() {
+        CONT_ACCESS_PANICS.store(0, Ordering::Relaxed);
+        let ca = ContAccess::new(42);
         let assert_fails = |fails: u8| {
-            let _ = cont.get_ref();
-            cont.write(0x42);
+            ca.copy(); // test everything
+            ca.write(15);
+            ca.btemap(|_| {});
             assert_eq!(CONT_ACCESS_PANICS.load(Ordering::Relaxed), fails);
         };
 
         assert_fails(0);
-        unsafe { INT_HANDLER_COUNT += 1 };
-        assert_fails(2);
-        unsafe { INT_HANDLER_COUNT -= 1 };
-        assert_fails(2);
+        INTERRUPT_DEPTH.fetch_add(1, Ordering::Relaxed);
+        assert_fails(3);
+        INTERRUPT_DEPTH.fetch_sub(1, Ordering::Relaxed);
+        assert_fails(3);
     }
 
-    /// Tests that references obtained from [`ContAccess::get_ref`] update after [`ContAccess::write`]s.
+    /// Tests that [`ContAccess::check_access`] panics when accessing it
+    /// inside  of it's [`ContAccess::btemap`].
     #[test_case]
-    fn refs_update() {
-        let cont = ContAccess::new(0x42);
-        let cont_ref = cont.get_ref();
-        assert_eq!(*cont_ref, 0x42);
-        cont.write(87);
-        assert_eq!(*cont_ref, 87);
-        assert_eq!(cont_ref, cont.get_ref())
+    fn badbug_on_locked() {
+        CONT_ACCESS_PANICS.store(0, Ordering::Relaxed);
+        let ca = ContAccess::new(42);
+        ca.btemap(|_| ca.copy());
+        ca.btemap(|v| ca.write(*v));
+        assert_eq!(CONT_ACCESS_PANICS.load(Ordering::Relaxed), 2);
+    }
+
+    /// Tests that [`ContAccess::btemap`], [`ContAccess::copy`] and
+    /// [`ContAccess::write`] all work correctly.
+    #[test_case]
+    fn everything_works_fine() {
+        let ca = ContAccess::new(42);
+        assert_eq!(ca.copy(), 42);
+        ca.btemap(|ans| assert_eq!(*ans, 42));
+        ca.write(87);
+        assert_eq!(ca.copy(), 87);
+        assert_eq!(ca.btemap(|ans| *ans), 87);
     }
 }